move code to constants.py and helpers.py
